@@ -10,8 +10,10 @@ const PRELOADED_DEPENDENCIES = [
   "react-dom",
   "@emotion/css",
   "@emotion/styled",
-  "@mantine/core",
-  "@mantine/prism",
+  "@kivra/toybox > @mantine/core",
+  "@kivra/toybox > @mantine/code-highlight",
+  "@kivra/toybox > @mantine/hooks",
+  "@tabler/icons-react",
   "mobx",
   "mobx-react-lite",
   "react-router-dom",
