@@ -1,5 +1,11 @@
 import styled from "@emotion/styled";
-import { Center, Title, Button, Collapse } from "@mantine/core";
+import {
+  Button,
+  Center,
+  Title,
+  Tooltip,
+  useMantineColorScheme,
+} from "@mantine/core";
 import { observer } from "mobx-react-lite";
 import { Story } from "../../../types";
 import { Markdown } from "./atom/Markdown";
@@ -7,20 +13,20 @@ import { ActionOutput } from "./ActionOutput";
 import { CodeTemplate } from "./CodeTemplate";
 import { Controls } from "./control/Controls";
 import { useControl } from "./control/useControl";
-import { useIsDarkMode } from "./control/useIsDarkMode";
 import { DefaultComponentWrapper } from "./DefaultComponentWrapper";
 import { useActionOutput } from "./useActionOutput";
-import { useState } from "react";
-import { ArrowDownIcon } from "../../icons/ArrowDownIcon";
+import { IconAccessible } from "@tabler/icons-react";
+import { useRef } from "react";
 
 export const StoryComponent = observer(({ story }: { story: Story }) => {
   const controls = useControl();
-  const darkMode = useIsDarkMode();
+  const darkMode = useMantineColorScheme();
   const { outputs, action } = useActionOutput();
+  const resetPointRef = useRef<HTMLButtonElement>(null);
 
   const renderdStory = story.render && (
     <DefaultComponentWrapper
-      colorScheme={darkMode.isDarkMode ? "dark" : "light"}
+      colorScheme={darkMode.colorScheme === "dark" ? "dark" : "light"}
     >
       {story.render(controls, action)}
     </DefaultComponentWrapper>
@@ -32,20 +38,15 @@ export const StoryComponent = observer(({ story }: { story: Story }) => {
 
   const sectionId = encodeURIComponent(story?.name ?? "");
 
-  const [opened, setOpen] = useState(false);
-
-  const toggleCodeSnippet = () => {
-    setOpen((prevOpened) => !prevOpened);
-  };
-
-  const handleClick = () => {
-    toggleCodeSnippet();
-  };
-
   let codeTemplateSource = null;
+
   if (story.codeTemplate) {
     codeTemplateSource = story.codeTemplate.toString().trim();
   }
+
+  const handleButtonClick = () => {
+    resetPointRef.current?.focus();
+  };
 
   return (
     <Wrapper>
@@ -53,68 +54,53 @@ export const StoryComponent = observer(({ story }: { story: Story }) => {
         <StoryTitle order={3}>{story.name}</StoryTitle>
       </SectionAnchor>
       {story.information && (
-        <div style={{ marginBottom: 32 }}>
+        <MarkdownGlobalStyles>
           <Markdown children={story.information} />
-        </div>
+        </MarkdownGlobalStyles>
       )}
       {renderdStory && (
         <>
           <Configurator>
-            <Prewview
-              style={{ background: darkMode.isDarkMode ? "#000" : "#fff" }}
+            <Preview
+              variant={story.variant}
+              hideControls={story.hideControls}
+              codeTemplate={Boolean(story.codeTemplate)}
             >
-              <ResetPoint tabIndex={-1} id={`reset-${sectionId}`} />
+              <ResetFocusButton>
+                <Tooltip withArrow arrowPosition="center" label="Reset focus">
+                  <Button
+                    variant="transparent"
+                    color="var(--text-secondary)"
+                    onClick={handleButtonClick}
+                    size="sm"
+                    leftSection={<IconAccessible size={24} />}
+                    style={{
+                      padding: "0px",
+                      marginBottom: "4px",
+                      display: "flex",
+                    }}
+                  ></Button>
+                </Tooltip>
+              </ResetFocusButton>
+              <ResetPoint tabIndex={-1} ref={resetPointRef} />
               {story.center ? (
                 <Center style={{ height: "100%" }}>{renderdStory}</Center>
               ) : (
                 <div>{renderdStory}</div>
               )}
-            </Prewview>
-            <ControlsWrapper>
-              <Controls
-                sectionId={sectionId}
-                controls={controls}
-                darkMode={darkMode}
-              />
-            </ControlsWrapper>
+            </Preview>
+            {!story.hideControls && (
+              <ControlsWrapper>
+                <Controls sectionId={sectionId} controls={controls} />
+              </ControlsWrapper>
+            )}
           </Configurator>
-          {story.codeTemplate &&
-            (story.codeTemplateExpanded ? (
-              <>
-                <Button
-                  leftIcon={<ArrowDownIcon />}
-                  variant="default"
-                  onClick={handleClick}
-                  styles={() => ({
-                    root: {
-                      color: "gray",
-                      margin: "20px 0",
-                    },
-                    leftIcon: {
-                      transform: opened ? "rotate(180deg)" : "rotate(0deg)",
-                      transition: "transform 0.3s ease-in-out",
-                    },
-                  })}
-                >
-                  {opened ? "Hide code example" : "Show code example"}
-                </Button>
-                <Collapse
-                  in={opened}
-                  transitionDuration={300}
-                  transitionTimingFunction="linear"
-                >
-                  <CodeTemplate
-                    codeTemplate={story.codeTemplate}
-                    controls={controls}
-                  />
-                </Collapse>
-              </>
-            ) : (
-              <CodeTemplate
-                codeTemplate={story.codeTemplate}
-                controls={controls}
-              />
-            ))}
+          {story.codeTemplate && (
+            <CodeTemplate
+              codeTemplate={story.codeTemplate}
+              controls={controls}
+            />
+          )}
           <ActionOutput outputs={outputs} />
         </>
       )}
@@ -147,36 +133,68 @@ const StoryTitle = styled(Title)`
 
 const Wrapper = styled("div")({
   marginTop: "12px",
-  marginBottom: "36px",
+  marginBottom: "48px",
 });
 
 const Configurator = styled("div")({
-  border: "1px solid #f1f3f5",
   display: "flex",
   maxWidth: "100%",
-  borderTopLeftRadius: "4px",
-  borderTopRightRadius: "4px",
   "@media (max-width: 960px)": {
     flexDirection: "column",
   },
 });
 
-const Prewview = styled.div`
-  position: relative;
-  flex: 1;
-  padding: 16px;
-  background: rgb(255, 255, 255);
-`;
+const Preview = styled.div<{
+  variant?: string;
+  hideControls?: boolean;
+  codeTemplate?: boolean;
+}>(({ variant, codeTemplate, hideControls }) => ({
+  position: "relative",
+  flex: "1",
+  padding: "16px",
+  background:
+    variant === "filled" ? "var(--surface-200)" : "var(--surface-100)",
+  borderTopLeftRadius: "12px",
+  borderTopRightRadius: hideControls ? "12px" : 0,
+  borderBottomLeftRadius: codeTemplate ? 0 : "12px",
+  borderBottomRightRadius: hideControls ? "12px" : 0,
+  border: "1px solid var(--border)",
+  borderRight: hideControls ? 0 : undefined,
+  borderBottom: hideControls ? 0 : undefined,
+  minHeight: "270px",
+}));
 
 const ControlsWrapper = styled("div")({
   width: "250px",
+  borderTopRightRadius: "12px",
+  borderBottomRightRadius: "12px",
   padding: "16px",
   boxSizing: "border-box",
-  borderLeft: "1px solid #f1f3f5",
+  border: "1px solid var(--border)",
+  borderBottom: 0,
+  background: "var(--surface-100)",
   "@media (max-width: 960px)": {
     width: "100%",
     borderLeft: "none",
-    borderTop: "1px solid #f1f3f5",
+    borderTop: "1px solid var(--border)",
+  },
+});
+
+const ResetFocusButton = styled("div")({
+  position: "absolute",
+  bottom: 0,
+  right: 0,
+  zIndex: 1,
+  padding: "0px",
+});
+
+const MarkdownGlobalStyles = styled("div")({
+  "p, h1, h2, h3, h4, h5, ul, ol, li": {
+    maxWidth: "860px",
+  },
+
+  p: {
+    marginBottom: "16px",
   },
 });
 
