@@ -1,7 +1,7 @@
-import { Group, ColorSwatch } from '@mantine/core';
-import { action } from 'mobx';
-import { observer } from 'mobx-react-lite';
-import { ColorControl as ColorControlType } from './useControl';
+import { Group, ColorSwatch } from "@mantine/core";
+import { action } from "mobx";
+import { observer } from "mobx-react-lite";
+import { ColorControl as ColorControlType } from "./useControl";
 
 interface Props {
   control: ColorControlType;
@@ -9,8 +9,8 @@ interface Props {
 
 export const ColorControl = observer(({ control }: Props) => {
   return (
-    <Group position="left" spacing="xs">
-      {control.options.map(c => {
+    <Group>
+      {control.options.map((c) => {
         return (
           <ColorSwatch
             key={c}
@@ -18,9 +18,9 @@ export const ColorControl = observer(({ control }: Props) => {
             color={c}
             onClick={action(() => (control.value = c))}
             style={{
-              color: '#fff',
-              cursor: 'pointer',
-              border: control.value === c ? '5px solid black' : 'none',
+              color: "#fff",
+              cursor: "pointer",
+              border: control.value === c ? "5px solid black" : "none",
             }}
           />
         );
