@@ -1,14 +1,16 @@
-import type { Config } from './src/public';
+import type { Config } from "./src/public";
 
 const config: Config = {
   eagerLoading: false,
-  storyPath: 'src/webapp',
+  storyPath: "src/webapp",
   wrapperComponent: {
-    path: '.toybox/Wrapper.tsx',
-    componentName: 'Wrapper'
+    path: ".toybox/Wrapper.tsx",
+    componentName: "Wrapper",
   },
-  title: 'Toybox',
-  emojiIcon: '🐒',
-}
+  title: "Toybox",
+  emojiIcon: "🐒",
+  npmPackageName: "@kivra/toybox",
+  startpageComponentPath: "startpage.story.md",
+};
 
 module.exports = config;
